backend/content_generation/requirements.txt
# MiniMax client dependencies
httpx>=0.27.0

# PDF text extraction
pypdf>=4.0.0

# DOCX text extraction (notes input)
python-docx>=1.1.0

# API server
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.0

# Video transcription (content pipeline)
openai-whisper>=20231117

# Word-level timestamps for subtitles
stable-ts>=2.16.0

# Reel production
Pillow>=10.0.0
PyMuPDF>=1.24.0

# AWS Transcribe (optional, for --transcriber aws)
boto3>=1.34.0
